requirements/requirements.txt
Django==4.2.5
psycopg2
django-extensions
djangorestframework==3.14.0
Pillow
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0